+++ modules/base/rendering/grids/renderableradialgrid.cpp
@@ -151,14 +151,14 @@ RenderableRadialGrid::RenderableRadialGrid(const ghoul::Dictionary& dictionary)
     registerUpdateRenderBinFromOpacity();
 
     if (dictionary.hasKey(GridColorInfo.identifier)) {
-        _gridColor = dictionary.value<glm::vec3>(GridColorInfo.identifier);
+        _gridColor = dictionary.value<glm::dvec3>(GridColorInfo.identifier);
     }
     _gridColor.setViewOption(properties::Property::ViewOptions::Color);
     addProperty(_gridColor);
 
     if (dictionary.hasKey(GridSegmentsInfo.identifier)) {
         _gridSegments = static_cast<glm::ivec2>(
-            dictionary.value<glm::vec2>(GridSegmentsInfo.identifier)
+            dictionary.value<glm::dvec2>(GridSegmentsInfo.identifier)
         );
     }
     _gridSegments.onChange([&]() { _gridIsDirty = true; });
