+++ modules/cefwebgui/cefwebguimodule.cpp
@@ -196,8 +196,10 @@ void CefWebGuiModule::internalInitialize(const ghoul::Dictionary& configuration)
         }
     );
 
-    if (configuration.hasValue<float>(GuiScaleInfo.identifier)) {
-        _guiScale = configuration.value<float>(GuiScaleInfo.identifier);
+    if (configuration.hasValue<double>(GuiScaleInfo.identifier)) {
+        _guiScale = static_cast<float>(
+            configuration.value<double>(GuiScaleInfo.identifier)
+        );
     }
 
     _enabled = configuration.hasValue<bool>(EnabledInfo.identifier) &&
